requirements.txt
alembic==0.9.7
astroid==1.6.1
Authlib==0.5.1
awscli==1.11.53
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
boto==2.48.0
boto3==1.5.32
botocore==1.8.46
bz2file==0.98
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.5.0
cycler==0.10.0
dask==0.12.0
decorator==4.0.10
docutils==0.14
dotenv==0.0.5
enum34==1.1.6
flake8==3.5.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
Flask-OAuthlib==0.9.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
futures==3.2.0
gensim==3.3.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
inflect==0.2.5
isort==4.3.4
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
joblib==0.11
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==0.23
matplotlib==1.5.3
mccabe==0.6.1
networkx==1.11
nltk==3.2.5
nose==1.3.7
numpy==1.14.0
oauth2==1.9.0.post1
oauthlib==2.0.6
pandas==0.22.0
pathlib==1.0.1
pew==1.1.2
Pillow==3.4.2
pipenv==9.1.0
psycopg2==2.7.3.2
pyasn1==0.2.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pylint==1.8.2
pymongo==3.6.1
pyparsing==2.1.10
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2016.7
PyYAML==3.12
rake-nltk==1.0.1
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.12.3
scikit-learn==0.18.1
scikit-surprise==1.0.5
scipy==1.0.0
shutilwhich==1.1.0
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.6
sqlacodegen==1.1.6
SQLAlchemy==1.0.14
sqlautocode==0.7
toolz==0.8.0
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
